Deprecate Cinder Hosts API/Extension
This was added a long long long long time ago, but it was never
fully implemented and is not used anywhere.  It might be worth
resurrecting and having, but it should probably be a first class
API rather than an extension, and if nothing else it would be good
if we want this to create a spec and have a real plan on it's
implementation and usage.

Bottom line, it's not used anywhere and the implementation is not
complete.  We could probably remove it safely, but let's deprecate
it and fast track removal and possible replacement next cycle.

Change-Id: I1a1920d141c8c32a8fb30bc6f73e955a1a1c5150