Stop a connections leak when freeing an SSHPool
In SSHPool when we free the SSHpool add a destructor
that closes all the connections in the SSHPool.

In the SSHPool when we resize to make the max_size smaller
that current_size or otherwise end up with a
current_size being greater than the max_size, we now close
the connection by adding a put method that closes the
connection before discarding the connection.

Added UT

Change-Id: I655458e39676e18b2dd8efe9dbc66544b4260e61
Closes-bug: 1748558