Stop unnecessarily querying storage for stats
Refreshing the stats is a costly operation for most drivers, but we are
still calling it on 3 manager actions that don't require it:

- delete_volume
- retype
- delete_group

This is probably due to an unfortunately naming of the method, because
`publish_service_capabilities` seem to indicate that we'll only be
publishing this information, but the method is also refreshing the stats
from the driver.

This patch removes all unnecessary calls to
`publish_service_capabilities`.

Change-Id: If2ae4abfa0e961607591239dabce8e539994a7da
Closes-Bug: #1750885