Fix allocated_capacity_gb race on create volume
Because of the stat reporting mechanism used in Cinder and because we
don't update the allocated_capacity_gb in the volume service until the
volume's creation has been successfully completed we have race
conditions between the scheduler's value and the volume's value that may
induce the scheduler to provision more volumes than what it should
allow.

This patches fixes this by assuming that the create volume operation on
the volume service will be successful, and thus increases the
allocated_capacity_gb value when the operation starts and reverts the
increase if we reschedule the creation.

Closes-bug: #1751057
Change-Id: Ia321a08d102a9a4f535e46158764cb7f9bc81a9d
(cherry picked from commit bf6df89a9e2e5240610995a9ca80388365ac1700)