Parse endpoint arguments in SolidFire
The SolidFire driver was assuming the same login/password and port
settings when building endpoint info for replication targets.  This
was true even if the config specified these fields as being different
in the config entry for the replication device.

This change updates the private build_endpoint method to actually
read any kwargs passed in to it and use them.

Change-Id: I8f95c891c9ee64fb8dd6b541be15fff539e2ab21
Closes-Bug: #1751917