OVO: support query for disabled services
Add a additional parameter to get_by_host_and_topic() to query for
disabled (or not) services.  This defaults to False so remains
transparent to existing callers.

This is needed for the coming replication promotion patch where services
are required to be disable prior to promotion.  If this is the case, I
must be able to access the disabled services in order to correctly
perform the promotion operation.

Change-Id: I8fc93c55546e6d4a582beaea9caebcbc9f18d98b