Replace assert to raise AssertionError
This patch is a fix for the issue B101 assert_used found by bandit
scanner.
As assert functionality could be turned off by the -O option for python
interpreter, assert statements were replaced to raise AssertionError.

The main reasoning for that is rest code after assert could be not
ready for values that assert filters.

Change-Id: I16d9c228f0f8aadae2044ce8fb95d7738970b0d5