NEC driver: Disallow access from the source node after live-migration.
NEC driver deletes access permission from nodes in the method
remove_export().
However, the method remove_export() is not called after live-migration.
After live-migration, access permission of the live-migration source
node is not deleted.

This patch adds the step to delete access permission from the node
in the method terminate_connection().
The terminate_connection() is called after live-migration and
access permission of the source node is deleted.

Change-Id: Id605eb6ff61cd1d382bf82173e6d65b986dda398
Closes-Bug: #1753375