VMAX driver - Replication failover performance improvement
Currently, failover of synchronous volumes on a VMAX backend can take a
long time as each volume is failed over consecutively. This patch
rectifies the issue by creating an internal group to facilitate the use
of group API's in the backend instead of volume API's, meaning the
volumes can failed-over concurrently.

Change-Id: Ie805c680662c160b77afcef47a3e0443bf1b15ef
Closes-bug: 1747894