Document more review processes
The Cinder team has some code review policies that are
generally known but not documented anywhere.  This patch
provides a place for us to document those reiview
practices that may not be known to everyone.

I have started by documenting the new process proposed at
the Dublin PTG for targeting milestones.  I also add
documentation to formalize the fact that we should not
merge vendor patches without their CI passing on the
patch.

Change-Id: Ie3bb23dcb6299f15ae69e12095c1aa9640565c33