Fix API reference attach_mode parameter default for attachments API
The default attach_mode for volume attachments is "rw", as can
be seen from looking at the VolumeAttachment.finish_attach method
and VolumeManager.attachment_update method. The REST API itself
does not set or return a default attach_mode, it's all set in the
data model when the attachment is 'completed' by assigning a host
connector.

Change-Id: I1e41e93bd534d830311a653eb16fef89a2d8431a