Dell EMC SC: ISCSI initialize_connection fixes.
Driver was not honoring the server in ISCSI and returning connection
information for all servers (IQNs) connected to a volume.

If multiple servers are mapped to a volume the driver was only returning
the first server for a given portal.

Closes-Bug: #1756914
Change-Id: If65780000fa8f59f663b00b88584cf289effa590
(cherry picked from commit 96b77e0aa62790320d16f7229aaeeb650fa875b6)