Adding Glance method for listing image members and unit test
This is necessary for checking from Cinder's side whether the current
project has access to an image with 'shared' visibility.  Often a
clone_image request will arrive at a driver with Glance pointed at the
same backend and within Glance the image is set to 'shared' visibility.
Currently the driver will have to call Glance directly to determine if
it should continue cloning the image, or return a permissions error.

Change-Id: Idb9641e136bfc38bbfa881452053056e74bd0553