Sync snapshot's encryption_key_id with volume's value
Sync the snapshot's encryption_key_id with the volume's value just
prior to updating the snapshot's DB entry. This ensures the snapshot
DB entry isn't out of date in case the volume's encryption_key_id
changed while the snapshot was in flight. The encryption_key_id will
change when keys based on the ConfKeyManager are migrated to Barbican.

Closes-Bug: #1756139
Change-Id: I65abb8dd17de7633828b731d1cf1f2321a6f3e5b