Avoid second restart on offline upgrades
On offline upgrades, due to the rolling upgrade mechanism we need to
restart the cinder services twice to complete the upgrade, just like in
the rolling upgrade case.

The current offline upgrade process is:

- Stop cinder services
- Upgrade the cinder nodes
- Sync your DB
- Start the cinder services
- Restart all the cinder services

This second restart creates a bad user experience and it should not be
necessary on an offline upgrade, so this patch adds a new optional
parameter -called "--bump-versions"- to the cinder-manage db sync
command that allows us to skip the restart of the services.

Closes-Bug: #1756321
Change-Id: I1b58c637f6b2187a78c9c00a6c4933335439ad6f
(cherry picked from commit 3cd2ebd3759c76fdf5a292e612127094c7aa2b17)