Fix: Incorrect replication status on clusters
When running a replicated service in a cluster, the replication_status
field in the Cluster DB table doesn't get updated in some cases, making
the cluster look like it's non replicated:

- When starting a new clustered replicated service cluster.
- When enabling replication on an already clustered service.

The reason for that is that the service and cluster creation happens on
the Service's __init__ method, while the replication status isn't known
until later, when the start method in the service calls the manager's
init_host_with_rpc, where the manager updates the service DB entry.

This patch adds an update for the replication field to the cluster table
in that same place, making sure we have the right data.

Closes-Bug: #1758024
Change-Id: I0d2fded62372a2fc5f49538133c5b66694dccc77
(cherry picked from commit 43b1209f2a8accf7d2744da3c7a552e92d8a55be)