Remove inappropriate directory space check
Removes an inappropriate directory space check when copying an image to
a volume. In the general case, "dest" is a block device, so it's not
appropriate to check for space in the directory containing "dest". One
exception may be the NFS driver, where dest actually is a file, but that
should be handled elsewhere (if necessary). This check has always been
inappropriate, but was silently ineffective due to a unit-comparison bug
that was recently fixed.

Change-Id: Ie5eba32f92eb1740554bfc44baf612e326d6fde5
Closes-Bug: 1756425