Reduce scope of the lock for image volume cache
Refactor the code that creates a volume from a downloaded glance image
to minimize the scope of the lock that prevents multiple entries in the
volume image cache. Now the lock serializes only the portion of the
code that causes the cache entry to be created. Locking is minimized
when the volume is already cached, or when the volume won't be cached.

Closes-Bug: #1758414
Change-Id: I547fb4bcdd4783225b8ca96d157c61ca3bcf4ef4
(cherry picked from commit a351cd01c0bf937e82efc3e1256da479be24cfcf)
(cherry picked from commit bc05efc79aaedd06308f7f173df09fd6a23aa2b3)