Unity: fail to detach lun when auto zone enabled
When the lun is the last one attached to the host and it is detached,
the attribute `host.host_luns` is None. Use `not` in if statement which
is pythonic in checking None and empty list of `host.host_luns`.

Change-Id: I5d8f63f9a78d001bc699af538f586aab8d73a5cf
Closes-bug: #1759175
(cherry picked from commit c816be897e7ab1c95b38979b2cc94ecf179e44e7)