Adding image membership check to _is_image_available
This adds a condition for valid membership where the image visibility is
'shared', the current project matches a member_id of the image and that
member's status is 'accepted'.  If all of these conditions are true,
then the current project has access to the image.  This check is
performed after other checks since it requires querying for membership
information from Glance so it will only affect membership checks that
were likely to fail already (due to having failed all previous checks).

Change-Id: Ifdcf7c5eab435b0dc40dd3650b61ebfef1e5579e