[SAP] implement size expansion when creating volume from template-based snapshot
[SAP] Fix create from snapshot with larger size

This patch fixes a problem when creating a volume from snapshot
when the size of the volume is larger than the original snapshot.
The problem was caused by python creating a float value from the
size calculation, and the vmware api not being able to process the
size float.  The value needs to be a whole number/int.

Change-Id: Ib84283511ee9fedfe08e0d24be49f9f239325432
This patch fixes some pep8 checks not covered during queens.