Clean up doc configuration
Remove some stale settings and get rid of setting version and
release. Those settings are now taken care of by openstackdocstheme
and do not need to be explicitly set, and since docs jobs no
longer install the project in order to generate docs, we are not
able to load those values from the python modules anyway.

Change-Id: I1a9a48538ca5c5e22e02f34c837e8fc783aeb2e9