Move MV history under API Microversion contrib docs
The REST API Version History was being listed in the top level
TOC for the contributor guide, causing each microversion to show
up as a separate item in the table of contents. This looks a odd
and causes a lot of scrolling to get to some of the contributor
information.

This moves the version history under the API Microversions
section so all microversion information is within on section of
the guide.

Change-Id: I74232b52088a395b618e07902427889200c2fdae