NetApp ONTAP: Set new sub-lun clone limit for ONTAP driver
Until ONTAP v8.3, the driver used to perform sub-lun clone of up to 64GB
per operation, but as of v9.1, ONTAP handles only 128MB per operation.
This new limit is causing error on volume extending operations, which
triggers sub-lun cloning when resizing exceeds max-resize-geometry.
For instance, if a volume is extended from 1GB to 64GB, the operation
fails because it triggers a single sub-lun cloning of 1GB.

This fix sets the limit used by sub-lun cloning to 128MB, therefore
if a volume is extended from 1GB to 64GB, it no longer fails, and
instead it will trigger 8 sub-lun cloning operations of 128MB.

Change-Id: Ib294cbdbdfe19d34a7702dafec2c8d29136a2f25
Closes-Bug: #1762424