ZFSSA accept pool stats when cluster node stripped
When a node in a ZFSSA cluster fails, storage pools and network
interfaces associated with it get "taken over" by the other node. In
this scenario, the identity of the node does not match the "owner" of
the pool. Currently, the cinder volume drivers consider this to be a
failure. With this change, it will not be considered a failure, so long
as the cluster node is stripped. If the cluster is operating normally,
the node responding to API requests must be the node that owns the pool.

Change-Id: I073cf97bc65f93acd5732c9b7072c7130e6b248f
Closes-Bug: 1762777
(cherry picked from commit 28b9d2eeef69ac979c5f5e1160643f73faf157a4)