ZFSSA implement "Enhance iSCSI multipath support"
A typical ZFSSA iSCSI deployment involves a ZFSSA cluster with two
control heads, two storage pools, and two network interfaces. Under
normal operation, one of the pools and one of the network interfaces
would be owned by one control head, and the others by the other control
head. This provides a form of "active/active" operation.

Cinder would be configured with two backends - one pointing to the pool
and an iSCSI target (group) normally served by the first head, and the
other the second head.

In the event of a failure of one control head, the other head will
"takeover" its storage pool and network interface, allowing service to
continue (perhaps with reduced performance) until the problem head is
restored, at which time the resources would be "given back".

In the takeover scenario, a sendtargets iSCSI discovery to the portal
address normally associated with the failed head will be serviced by the
other head, and it will return BOTH targets. os-brick, when configured
to use multipath, interprets this as both targets being basically
equivalent, and will use either one of them to determine the
host/controller/target/LUN ID to use to find SCSI devices. When the same
LU number is assigned to LUNs in both pools that are presented to the
same compute node, it can cause a LUN belonging to one VM to to
connected to a different VM.

Whilst full multipath support is not implemented in the ZFSSA driver,
using multipath on compute nodes provides some utility, even with only a
single path (notably the I/O can be queued during a transient interruption
(cluster takeover, network switch reboot, etc.). To support this scenario,
whilst avoiding the above problem, this change implements "enhanced multipath
support" [1] to provide explicit target info to os-brick (but only
supporting a single path). This could be expanded to full multipath support
in the future.

[1] https://blueprints.launchpad.net/cinder/+spec/iscsi-multipath-enhancement

Change-Id: I6402a6bc2a0d51efd3b60d93c11c1a81e8a73269
Closes-Bug: 1762584
(cherry picked from commit a412e3279070140bfd4611d25cc6efc3b1b533c1)