VMAX driver - Attaches/detaches after host assisted retype failing
There is an issue after host-assisted retype where subsequent actions
on a retyped volume fail due to a volume ID mismatch. This patch
rectifies the issue by checking if the '_name_id' attribute of the
volume matches the volume identifier on the backend, should the volume
'id' attribute comparison fail.

Change-Id: I8badcbd99e25336b1a905b82959e6c3a7bfcc420
Closes-bug: 1763555