Improve API method logging
We have tracing in the wsgi code to log the method being called,
but since we are just formatting the function object into a string,
we end up with the less than friendly output of:

Calling method '<bound method VolumeController.delete of
<cinder.api.v3.volumes.VolumeController object at 0x7f9f14f86dd0>>'

This changes the logging to extract the actual name of the method
being called. It also removes some unnecessary string coersion.

Change-Id: I3a0a5973a1798a7fcf25c4288b9cbef1195f7015