NetApp ONTAP iSCSI: Force exception on online extend
The Netapp ONTAP iSCSI driver does not support online volume extend. It
may work if the requested size does not exceed the LUN max geometry,
otherwise it will require the LUN to be detached. In such case, the
backend currently detaches and leaves the volume in an inconsistent
state.

This patch forces the ONTAP iSCSI driver to raise an exception whenever
an online extend is requested and it detects it would exceed the LUN max
geometry.

Change-Id: Ie3dddbc05c6cd32e27168d68f4cb819364b0438c
Closes-Bug: #1712651