Add config table sphinx directive
In the past we had used an openstack-manuals tool to manually generate
config option tables that would then be included into driver config
documentation.

With the move of documentation in-tree and the deprecation and removal
of that tool, we have ended up with options that are no longer being
updated when drivers change, or maintainers are left manually updating
the existing tables.

This addes a sphinx extension to use a new config-table directive to
automatically pull in config options from the source so we no longer
need to perform any manual action to pick up changes.

Change-Id: I625fb96229001c326ed2400155e2d067279a400e