VMAX driver - Incorrect service level reporting
'Optimized' is being reported as a service level option for VMAX
All-Flash arrays, in the get_pools output, and it can be assigned as
the default pool if there is no service level assigned to a volume type.
This can cause errors, as Optimized is not a recognized service level on
All Flash arrays. This patch rectifies the issue.

Change-Id: I9365ae3b923eeb92ac00023b7040e934b6e2df57
Closes-bug: 1766633