Add reno to test-requirements
Reno was moved over to doc/requirements.txt when that was created
because it is needed for releasenotes builds. But with its removal
from the other requirements files, doing something like the
recommended "tox -e venv -- reno new xxx" to create a new release
note fails with reno missing from any venv's created since it was
removed.

This adds reno back in to test-requirements.txt so it is installed
when creating the venv, ummm... venv.

Change-Id: Id2e0325b0ce8e18bb9a65318fb3fc52777753536