GoodnessWeigher schedules non-type volumes
If the GoodnessWeigher is configured to be the only active weigher within
the scheduler, it cannot schedule a non-typed volumed for creation. This
patch adds a safety assignement to the weigher stats generation to allow
for scheduling untyped volumes.

This patch safely assigns a dictionary data-type to "volume_type" in the
event that:

    weight_properties = {
        "volume_type": None,
        etc..
    }

Volumes created with a volume-type will not cause an exception and will
be scheduled with the volume-type and back-end as configured.

Closes-Bug: 1766966

Change-Id: I7bc77ab02154aee80282e54b19a242bdea0a78ec
Signed-off-by: Eric M Gonzalez <eric@awnix.com>