Implement privsep boilerplate in cinder.
This includes implementing a first trivial example of how to use
privsep to run something as root, specifically the cgroup throttling
driver.

This code is modelled strongly on how nova has chosen to use privsep.
Consistency is probably good here, but it does imply that the cinder
team is ok with the decisons nova has made about implementation.

Change-Id: Ic401138a10a72cb4b976a1a6aba272cafcb40d8b