Fix google backup driver
Current backup driver is broken due to the upper constraints of
google-api-python-client which is set to 1.6.6, but after 1.6.0
parameter http and credentials are mutually exclusive, the driver is
also using the no longer preferred Google authentication library.

This patch fixes this by only using the http_proxy environmental
variable to set the proxying and adding support for google-auth and
using it whenever possible.

Closes-Bug: #1736569
Closes-Bug: #1769660
Change-Id: If1e34ea112b0cff328a58ec2fc9c3f5a5005f1c6