Remove os-image-create API extension
This was a dummy extension and only used by the v1 API. This is now a
standard part of the API and does not need any extra checks to see if it
should be enabled or not. This removes the dummy extensions and checks
in the code for its presence.

Change-Id: Ia0a3fa41aa06fbbe9d2ed67a1273226ee47429f9