Fix rbd driver json loading
It looks like rbd.py sometimes gets some bytes, not strings, probably
this happens with the latest version of Ceph, and that's an issue only
with Python 3. I'm not sure why, but what I know, is that it's completely
breaking Debian with Python 3 and ceph Luminous release.

Convert byte buffers coming from Ceph to strings.

Co-Authored-By: Eric Harney <eharney@redhat.com>
Change-Id: I5cdb4e766655a866699a18f1e367971fd1d310af
Closes-Bug: #1771552