Update author in latex sphinx doc builds
This commit updates the author in the latex build metadata in the sphinx
conf.py. This was set to Anso Labs, LLC from the initial fork out of
Nova in 2012. It's highly unlikely that an LLC acquired by Rackspace in
2011 is actually the author of the latex version of the docs so this
commit changes it to be "Cinder Contributors" to be accurate.

Change-Id: I5f1f7ea2f389783fe8b936d67d0ab99d759c4124