VMAX Driver - Incorrect SSL cert verification
SSL certificates are not being applied to the REST session as desired.
If a user sets SSL verification to True but does not specify a path
to the SSL cert, SSL is disabled instead of looking in the system
certificate bundle.

Closes-Bug #1772924

Change-Id: I20cb64c11e43c22558939e27708ec678c5723ca2
(cherry picked from commit 512fd07124ea7210a6653e519e964a898a31d406)