Fix handling of 'cinder_encryption_key_id' image metadata
The Cinder code that processes Glance image metadata
is a bit confused about whether this particular field
is a Glance property or metadata.

Since it isn't a defined a Glance property and is stored
in image metadata, ensure that Cinder also tracks it
metadata and not as a property.

This mismatch prior to this fix causes Cinder to create
volumes with the wrong encryption key when creating a
volume from an encrypted image, which results in an
unreadable volume.

Closes-Bug: #1764125
Change-Id: Ie5af3703eaa82d23b50127f611235d86e4104369