Change check_uptodate.sh to not call tox
Base nodepool images now no longer include tox. This causes a failure in
our pep8 jobs because the tools/config/check_uptodate.sh script had been
calling "tox -e genopts" to look for config file changes.

This is a partial backport of a change that switched our default
basepython to python3. That part of the change was removed from tox.ini,
so the only other part was an update to the check_uptodate.sh script
that refactored things to be more reliable, including the removal of
calling tox.

The part from the original commit included is:

There was also a difference in behavior with the genopts job
where it ends up trying to recreate the venv under which it
is currently running under, resulting in a corrupted venv
and a failure. This cleans up that script and changes it so
rather than a tox job calling a tox job it just runs the
script.

Change-Id: Ie79528b33070e6c60f9f5e1f2219084a2f0cb039
(cherry picked from commit 37510ac03478704bafba89e8c4f4e0e9041d661d)