Update macOS devref info
Mac OS X was renamed to macOS starting with Sierra in 2016. This
updates the development environment information to update the name
and note that unit tests and development are not fully supported
when using this OS.

Includes various doc formatting cleanup as well since I was making
changes to the file.

Change-Id: Ia62ac91add51450dc99c431e9f64732553495558