HPE3PAR: Fix pointing to backend in group failover
Issue: After group failover, subsequent operations
like attach & detach volume, which are part of that
failed over group points to primary backend instead
of secondary.

This patch fixes the above issue by setting up the
appropriate backend in case of subsequent operation
on volume, which are part of group which is failed
over.

Change-Id: I679b11317c91ad28cefdf995a8d6849dc71bc1c5
Closes-Bug: #1773069