Fix vmware disk type changed when volume type is eagerZeroedThick
Create a volume from an image, choose "eagerZeroedThick" volume_type.
But login in Vcenter, you will get a "lazyZeroedThick" volume. There
is no problem when creating a blank vmware volume.

The reason is: when creating a volume from image, if volueme_size
> image_size, the last step is to extend backing volume. In the method
extend_virtual_disk(), there is a parameter "eager_zero", which
determining if the free space is zeroed out. The param "eager_zero"
default value is "False", and not to be changed to "True" in creating
an eagerZeroedThick volume from an image. So the actuall volume type
is different with designated volume type.

We can add a paramter "disk_type" in _extend_backing() method,
determining which type disk to extend. Then pass "False" or "True" to
extend_virtual_disk() as "eager_zero".

Change-Id: I29978b69d21c2583da152f9fbf1264ada8f079e6
Closes-Bug: #1777430